requirements.txt
boto3==1.38.12
python-dotenv==1.0.1
openai==1.1.0
requests==2.31.0
mkdocs==1.6.3
